fix: cast isnt() comparison to numeric to match sanctioned_load_kw's type
Both tenants now provision cleanly (0054's fix confirmed working: "no
provisioning error logged" for both). This is the last remaining issue:
pgTAP's isnt(anyelement, anyelement, text) couldn't unify numeric against
a bare integer literal 999 — no matching overload for
(numeric, integer, unknown). Cast the literal to numeric.

+++ supabase/tests/rls/sandbox_isolation.test.sql
@@ -102,7 +102,7 @@ update service_connections set sanctioned_load_kw = 999 where dt_id = :'b_dt_id'
 reset role;
 select isnt(
   (select sanctioned_load_kw from service_connections where dt_id = :'b_dt_id'::uuid limit 1),
-  999,
+  999::numeric,
   'tenant A''s discom_officer''s write attempt on tenant B''s connection matched zero rows under RLS'
 );
 set local role authenticated;
